Started actual work
Started to work on the login page as well as the backend. Got the login page written Just have to styleize at somepoint. What im running with first is that I write the page and not do any styleizing yet. Just get the frame down and working then style later. ALso another Idea im running with is that everything is going to be locked behind the login page. Nothing can be accessed unless a user is authenticated. Cant test auth until I got other pages created. with some temp data to test against. Going to have to begin preperations for the API(s) that are going to be introduced at some point. Going to most likely be breaking that out into its own file as like an API hander and then just importing the information gather method.