Got some movement
Created the Registration Page. Got something basic for the homepage. the data that is displayed currently is still goign to be displayed, Its just going to be cleaned up and look nicer in the end. Besides just a general Listing, I have no idea what I want shown on the homepage. Going to figure that out later.